Moved common entities to another project
