--- WeatherDiaryApp/Common/Class1.cs
@@ -1,6 +1,0 @@
-﻿using System;
-
-namespace Common
-{
-    public class Class1 { }
-}
