+++ WeatherDiaryApp/Common/WeatherRecord.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace Server.Models
+namespace Common
 {
     public class WeatherRecord
     {
